+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.11.3](https://github.com/Headstorm/foundry-ui/compare/v1.11.2...v1.11.3) (2021-09-08)
+
+
+### Bug Fixes
+
+* **textinput:** enable display of character count without max length ([90844f3](https://github.com/Headstorm/foundry-ui/commit/90844f3db583ed109f39a9eaab41081b3399fb96)), closes [#295](https://github.com/Headstorm/foundry-ui/issues/295)
+* **textinput:** make uncontrolled input clearable ([4b1a4f0](https://github.com/Headstorm/foundry-ui/commit/4b1a4f0b038b60c62387a8583bc135e5e7e03fba)), closes [#295](https://github.com/Headstorm/foundry-ui/issues/295)
+* **textinput:** reset character count on clear ([abd36a5](https://github.com/Headstorm/foundry-ui/commit/abd36a59b027ddb47ece7b39b05bfac57a04977d)), closes [#295](https://github.com/Headstorm/foundry-ui/issues/295)
+
+## [1.11.2](https://github.com/Headstorm/foundry-ui/compare/v1.11.1...v1.11.2) (2021-09-07)
+
+
+### Bug Fixes
+
+* **button:** skeleton props object is now correctly spread on skeleton shimmer ([1fcf324](https://github.com/Headstorm/foundry-ui/commit/1fcf324055fc9667b168dcc054132657c51798b7)), closes [#322](https://github.com/Headstorm/foundry-ui/issues/322)
+
 ## [1.11.1](https://github.com/Headstorm/foundry-ui/compare/v1.11.0...v1.11.1) (2021-08-31)